encription scripts needed some fault tolerance and progress on hw03

+++ decrypt_hw.sh
@@ -1,4 +1,6 @@
 gpg-zip --output homework/my_solutions.tar --decrypt homework/my_solutions.tar.gpg
-rm -R homework/my_solutions.tar.gpg
-tar -xf homework/my_solutions.tar
-rm -R homework/my_solutions.tar
+if [ $? == 0 ]; then
+  rm -R homework/my_solutions.tar.gpg
+  tar -xf homework/my_solutions.tar
+  rm -R homework/my_solutions.tar
+fi
